CONN-BASE-0001: rename Connector.ts -> connector.ts (kebab-case CI gate)
The repo's 'Enforce Phase 2B Naming Standards' check requires kebab-case file
names; Connector.ts (PascalCase) was the lone violation. Rename to connector.ts
and update imports. No logic change; full suite 98/98 green.
